Refactored useHandleRequestError into useHandleResponseError for clarity
Also moved the logic for parsing out the error response into the composition handler so that we can just reject the Axios error straight on through.
