+++ src/components/players/PlayerRegistration.vue
@@ -43,7 +43,7 @@
 </template>
 
 <script>
-import useHandleRequestError from '/src/composites/useHandleRequestError.js'
+import useHandleResponseError from '/src/composites/useHandleResponseError.js'
 import PlayerBadge from '../shared/PlayerBadge.vue'
 import TextEditor from '../shared/TextEditor.vue'
 import TextInput from '../shared/TextInput.vue'
@@ -58,8 +58,8 @@ export default {
     TextInput,
   },
   setup () {
-    // Standard composite containing { toast, handleRequestError }
-    return useHandleRequestError()
+    // Standard composite containing { toast, handleResponseError }
+    return useHandleResponseError()
   },
   data: () => ({
     username: '',
@@ -96,7 +96,7 @@ export default {
       this.$store.dispatch('player/register', data).then(() => {
         this.toast.success(`Welcome to Ashes.live, ${this.username}!`)
         this.$router.push('/decks/mine/')
-      }).catch(this.handleRequestError)
+      }).catch(this.handleResponseError)
     },
   },
 }
